Move download into Manage Projects
